requirements.txt
# NerdX Bot Dependencies - Python 3.13 Compatible
# Core Flask and Web Framework
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.0
flask-cors>=6.0.0
flask-dance>=7.0.0
werkzeug>=3.0.0
gunicorn>=22.0.0

# Build tools (required for Python 3.13)
setuptools>=68.0.0
wheel>=0.41.0

# Database and ORM
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Authentication and Security
oauthlib>=3.2.0
pyjwt>=2.8.0
email-validator>=2.2.0
bcrypt

# AI and Machine Learning
google-generativeai>=0.8.0
google-genai
openai>=1.0.0
sympy>=1.12

# Vector Search and Embeddings (Free alternatives to Pinecone + OpenAI)
chromadb>=0.4.0  # Embedded vector database (stable on Windows)
sentence-transformers>=2.2.0  # Free sentence embeddings

# Animation Engine
manim>=0.18.0  # Math animation engine (requires ffmpeg + latex)

# Voice Features
openai-whisper>=20231117
edge-tts>=6.1.9  # High-quality free neural TTS
soundfile>=0.12.1
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Image Processing and OCR
pillow>=10.0.0
pytesseract>=0.3.10
imgbbpy>=0.1.0
pix2text[multilingual]>=1.1.0  # Free alternative to Mathpix OCR

# Document Processing
pypdf>=4.0.0
PyPDF2
reportlab>=4.0.0
python-docx

# Audio and Speech
speechrecognition>=3.10.0

# HTTP and API
requests>=2.31.0

# Payment Integration
paynow

# Additional Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Development and Testing (optional for production)
pytest>=7.4.0
pytest-cov>=4.1.0
